Make bookie client an interface
So that it can be mocked easily for testing.
This patch also containsa simple mock.
